upgrade many dependencies. Upgrade firebase functions to node-18. Change functions integration tests
