+++ _config.yml
@@ -15,8 +15,6 @@ include:
   - ".htaccess"
   - "node_modules/bootstrap/dist/js/bootstrap.bundle.min.js"
   - "node_modules/bootstrap/dist/js/bootstrap.bundle.min.js.map"
-  - "node_modules/cookieconsent/build/cookieconsent.min.css"
-  - "node_modules/cookieconsent/build/cookieconsent.min.js"
 highlighter: rouge
 incremental: true
 sass:
