--- _includes/scripts/analytics_privacy.html
@@ -1,79 +1,0 @@
-
-<!-- Begin Cookie Consent plugin by Silktide - http://silktide.com/cookieconsent -->
-<script defer type="text/javascript">
-  //Being denied isn't the same as not-consented-yet
-  var is_consented = document.cookie.indexOf('cookieconsent_status=allow') != -1;
-  var is_denied = document.cookie.indexOf('cookieconsent_status=dismiss') != -1;
-
-  var tracking = 'UA-48469340-1';
-  var disableString = 'ga-disable-'+tracking;
-  //We can prepare gtag function whatever the final choice is
-  //It will just never get sent if cookies are refused
-  window.dataLayer = window.dataLayer || [];
-  function gtag(){dataLayer.push(arguments);}
-  
-  function change_preferred_language(lang){
-    gtag('event', 'click', {event_category: 'lang', event_label: lang});
-    if(is_consented){
-      var d = new Date();
-      d.setTime(d.getTime() + (360*24*60*60*1000));
-      document.cookie = 'lang='+lang+'; expires='+d.toUTCString()+'; path=/;'
-    }
-  }
-
-  if (is_denied) {
-    console.log("disabling gtag because cookieconsent_status is set to dismiss");
-    window[disableString] = true;
-  }else if(is_consented){
-    gtag('js', new Date());
-    gtag('config', tracking);
-  }
-
-  //cookieconsent does automatically store user response and call this callback
-  function registerGa(status){
-    // if cookie wasn't set we wait for consent to call gtag
-    if(!is_consented && status== "allow" && this.hasConsented()){
-      is_consented = true;
-      gtag('js', new Date());
-      gtag('config', tracking);
-    }
-  }
-
-  //Always create cookieconsent to give revoke option
-  var ccStyle = document.createElement("link");
-  ccStyle.rel = "stylesheet";
-  ccStyle.href = "/css/cookieconsent.css";
-  document.head.appendChild(ccStyle);
-  var cc = document.createElement("SCRIPT");
-  cc.src = '/node_modules/cookieconsent/build/cookieconsent.min.js';
-  cc.async = true;
-  cc.onload = function(){
-    console.log("cookieconsent loaded")
-    window.cookieconsent.initialise({
-      type: "opt-in",
-      container: document.body,
-      revokable: true,
-      content:{
-        message: "{{site.data.strings.cookieconsent[page.lang]}}",
-        link: "{{site.data.strings.read-more[page.lang]}}",
-        href: '/{{page.lang}}/about/privacy',
-        dismiss: "&times;",
-        allow: "&#x2713;",
-      },
-      palette:{
-        "popup": {
-          "background": "var(--dark)",
-          "text": "var(--light)"
-        },
-        "button": {
-          "background": "var(--info)",
-          "text": "#ffffff"
-        }
-      },
-      onStatusChange: registerGa,
-      onInitialise: registerGa
-    });
-  }
-  document.head.appendChild(cc);
-</script>
-<script async defer src="https://www.googletagmanager.com/gtag/js?id=UA-48469340-1"></script>
