--- css/cookieconsent.scss
@@ -1,43 +1,0 @@
----
-# this ensures Jekyll reads the file to be transformed into CSS later
-# only Main files contain this front matter, not partials.
----
-@import "variables";
-@import "/node_modules/cookieconsent/build/cookieconsent.min.css";
-.cc-window{
-  &.cc-banner, &.cc-revoke{
-    background: $blue;
-    color: #fff;
-    font-family: "Helvetica Neue Light", "HelveticaNeue-Light", "Helvetica Neue", Calibri, Helvetica, Arial;
-    box-sizing: border-box;
-    left:auto;
-    bottom: 1em;
-    right: 1em;
-    width:auto;
-    max-width:50%;
-    border-radius: 4px;
-    padding: 0 0 0 1em;
-  }
-}
-
-
-.cc-compliance{
-  padding-left: 15px;
-}
-
-.cc-banner .cc-btn{
-  font-size: 32px;
-  line-height: 21px;
-  vertical-align: middle;
-  font-weight: bold;
-  &:last-child{
-    min-width: 51px;
-  }
-}
-.cc-btn.cc-deny{
-  color:$dark-blue;
-}
-.cc-btn:hover{
-  color: $white!important;
-  text-decoration: none !important;
-}
