+++ firebase.json
@@ -9,11 +9,11 @@
     "rewrites": [
       {
         "source": "/",
-        "function": "api_calls"
+        "function": "http_calls"
       },
       {
         "source": "/api/v1/**",
-        "function": "api_calls"
+        "function": "http_calls"
       }
     ],
     "redirects": [
@@ -51,5 +51,13 @@
   "firestore": {
     "rules": "firestore.rules",
     "indexes": "firestore.indexes.json"
+  },
+  "emulators": {
+    "firestore": {
+      "port": "8080"
+    },
+    "functions": {
+      "port": "5001"
+    }
   }
 }
