+++ functions/__mocks__/firebase-functions/logger.js
@@ -1,5 +1,4 @@
 
-
 module.exports = {
   info: jest.fn(),
   warn: jest.fn(),
