+++ functions/package.json
@@ -3,32 +3,32 @@
   "description": "Cloud Functions for Firebase",
   "scripts": {
     "lint": "eslint .",
-    "serve": "firebase emulators:start --only functions",
+    "serve": "firebase emulators:start --only hosting,functions,firestore",
     "shell": "firebase functions:shell",
     "start": "npm run shell",
-    "test": "firebase emulators:exec jest",
-    "watch": "firebase emulators:exec 'jest --watch'",
+    "test": "firebase emulators:exec 'jest --runInBand'",
+    "watch": "firebase emulators:exec 'jest --watch --runInBand'",
     "deploy": "firebase deploy --only functions",
     "logs": "firebase functions:log"
   },
   "engines": {
-    "node": "16"
+    "node": "18"
   },
   "main": "index.js",
   "jest": {
     "testEnvironment": "node"
   },
   "dependencies": {
-    "body-parser": "^1.19.0",
+    "body-parser": "^1.20.2",
     "express-rate-limit": "^5.1.3",
-    "firebase-admin": "^8.10.0",
-    "firebase-functions": "^3.20.1",
-    "node-fetch": "^2.6.1"
+    "firebase-admin": "^11.9.0",
+    "firebase-functions": "^4.4.0",
+    "node-fetch": "^2.6.7"
   },
   "devDependencies": {
-    "eslint": "^5.12.0",
-    "eslint-plugin-promise": "^4.0.1",
-    "jest": "^26.5.0",
+    "eslint": "^8.42.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "jest": "^29.5.0",
     "supertest": "^5.0.0"
   },
   "private": true
