+++ package.json
@@ -16,15 +16,14 @@
     "dist/"
   ],
   "devDependencies": {
-    "@popperjs/core": "^2.11.6",
+    "@popperjs/core": "^2.11.8",
     "bootstrap": "^5.2.0",
-    "chai": "^4.3.6",
-    "cookieconsent": "^3.0.6",
-    "faker": "^4.1.0",
-    "finalhandler": "^1.1.1",
-    "firebase-tools": "^11.8.0",
+    "chai": "^4.3.7",
+    "faker": "^5.5.3",
+    "finalhandler": "^1.2.0",
+    "firebase-tools": "^12.3.1",
     "js-yaml": "^4.1.0",
-    "mocha": "^10.0.0",
+    "mocha": "^10.2.0",
     "puppeteer": "^17.0.0",
     "serve-static": "^1.15.0",
     "xml-parser": "^1.2.1"
